Rename action `deleteUsers` for consistency
