+++ hasura/metadata/actions.yaml
@@ -133,10 +133,10 @@ actions:
           value_from_env: HASURA_EVENT_SECRET
     permissions:
       - role: user
-  - name: deleteUserBulk
+  - name: deleteUsers
     definition:
       kind: synchronous
-      handler: '{{HASURA_API_BASE_URL}}/actions/actionManager?action_name=deleteUserBulk'
+      handler: '{{HASURA_API_BASE_URL}}/actions/actionManager?action_name=deleteUsers'
       headers:
         - name: verification_key
           value_from_env: HASURA_EVENT_SECRET
@@ -353,6 +353,7 @@ custom_types:
     - name: DeleteDiscordUserInput
     - name: LinkDiscordCircleInput
     - name: DeleteUserBulkInput
+    - name: DeleteUsersInput
     - name: UpdateProfileInput
   objects:
     - name: CreateCircleResponse
@@ -561,5 +562,6 @@ custom_types:
           type: object
     - name: LinkDiscordCircleResponse
     - name: DeleteUserBulkResponse
+    - name: DeleteUsersResponse
   scalars:
     - name: timestamptz
